Refactor AbstractCanvas component and rename ActionButtons component
